+++ shinx121_debug/src/debug_common.c
@@ -57,9 +57,23 @@ void dbgui_input_short_ext(const char *label, s16 *value, const DbgUiInputIntOpt
     }
 }
 
+void dbgui_input_ushort_ext(const char *label, u16 *value, const DbgUiInputIntOptions *options) {
+    s32 value_ = *value;
+    if (dbgui_input_int_ext(label, &value_, options)) {
+        *value = (u16)value_;
+    }
+}
+
 void dbgui_input_uint(const char *label, u32 *value) {
     s32 value_ = *value;
     if (dbgui_input_int(label, &value_)) {
+        *value = (u32)value_;
+    }
+}
+
+void dbgui_input_uint_ext(const char *label, u32 *value, const DbgUiInputIntOptions *options) {
+    s32 value_ = *value;
+    if (dbgui_input_int_ext(label, &value_, options)) {
         *value = (u32)value_;
     }
 }